More TupleType fixing
Tuples (and UDTs) allow nulls for all fields, and their null is different from what individual types use for null. However, it does not make sense to have the individual type's null in a tuple and it appears user-supplied data cannot do so.
This changes the byte-comparable encoding to use the tuples' treatment of nulls instead of the type's.